Moved ControlOverrider to LIUtils
