+++ src/main/java/cn/academy/ability/api/ctrl/ClientController.java
@@ -28,10 +28,10 @@ import cn.academy.ability.api.event.PresetSwitchEvent;
 import cn.academy.ability.api.event.PresetUpdateEvent;
 import cn.academy.core.AcademyCraft;
 import cn.academy.core.ModuleCoreClient;
-import cn.academy.core.util.ControlOverrider;
 import cn.annoreg.core.Registrant;
 import cn.annoreg.mc.RegEventHandler;
 import cn.annoreg.mc.RegInit;
+import cn.liutils.util.client.ControlOverrider;
 import cn.liutils.util.helper.KeyHandler;
 import cn.liutils.util.helper.KeyManager;
 import cpw.mods.fml.common.eventhandler.SubscribeEvent;
