+++ src/main/java/cn/academy/terminal/client/TerminalUI.java
@@ -24,7 +24,6 @@ import cn.academy.core.ModuleCoreClient;
 import cn.academy.core.client.Resources;
 import cn.academy.core.client.sound.ACSounds;
 import cn.academy.core.registry.RegACKeyHandler;
-import cn.academy.core.util.ControlOverrider;
 import cn.academy.terminal.App;
 import cn.academy.terminal.AppEnvironment;
 import cn.academy.terminal.AppRegistry;
@@ -43,6 +42,7 @@ import cn.liutils.cgui.gui.component.TextBox;
 import cn.liutils.cgui.gui.event.FrameEvent;
 import cn.liutils.cgui.gui.event.FrameEvent.FrameEventHandler;
 import cn.liutils.cgui.loader.xml.CGUIDocLoader;
+import cn.liutils.util.client.ControlOverrider;
 import cn.liutils.util.generic.MathUtils;
 import cn.liutils.util.helper.GameTimer;
 import cn.liutils.util.helper.KeyHandler;
